LayoutTests/platform/win/fast/frames/frame-with-noresize-can-be-resized-after-setting-noResize-to-false-expected.txt



--------
Frame: 'testFrame'
--------


--------
Frame: 'results'
--------
This tests that a frame with a source code-specified noresize attribute can be resized after setting frame.noResize to false.

FAIL document.getElementById("testFrame").width should be 210. Was 200.
